pyproject.toml
[project]
name = "gemini-ollama-proxy"
description = "A lightweight proxy that lets you use Google's Gemini API through an Ollama-compatible interface."
version = "0.1.0"
requires-python = "~=3.11"
dependencies = [
  "fastapi~=0.115.13",
  "google-genai~=1.21.1",
  "pydantic~=2.11.5",
  "python-dotenv~=1.1.0",
  "uvicorn~=0.34.3",
]

[project.optional-dependencies]
dev = ["ruff"]

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[tool.ruff]
src = ["src"]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.format]
quote-style = "double"
